TweetSchedule関連の実装

+++ src/test/resources/schema.sql
@@ -57,7 +57,7 @@ CREATE INDEX idx_scheduled_tweets_delete_flag ON SCHEDULED_TWEETS (delete_flag);
 
 -- Create BATCH_HISTORY table
 CREATE TABLE IF NOT EXISTS BATCH_HISTORY (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    id INT AUTO_INCREMENT PRIMARY KEY,
     last_processed_tweet_id INT NOT NULL,
     processed_num INT NOT NULL,
     execution_start DATETIME(0) NOT NULL DEFAULT CURRENT_TIMESTAMP,
